requirements.txt
#torch==1.12.1
#torchvision
scikit-learn
numpy
tqdm
pandas
#transformers==4.21.2
Pillow==9.1.1
timm==0.6.7
clip @ git+https://github.com/openai/CLIP.git@b4ae44927b78d0093b556e3ce43cbdcff422017a
docarray==0.15.4
